+++ taxel-gui/CHANGELOG.md
@@ -42,3 +42,4 @@
   - Make error message clickable and jump to relevant fact
   - Copy diagnostic message on right click
   - Send report in test mode via env var `TEST_MARKER`
+  - Add modal to confirm terms of use and privacy notice